requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
dj-database-url==3.0.1
psycopg2-binary==2.9.11
pillow==11.3.0
tzdata==2025.2
python-dotenv==1.0.0
gunicorn==21.2.0
cloudinary==1.32.0
django-cloudinary-storage==0.3.0
